+++ NEWS.md
@@ -1,5 +1,10 @@
 # tidytable 0.6.1 (in development)
 
+#### Bug fixes
+* `if_all.()`/`if_any.()`
+  + No longer try to use `.by` columns (#225)
+  + Can use `n()` in filtering (#226)
+
 # tidytable 0.6.0
 
 #### New functions